+++ src/mailer.ts
@@ -17,10 +17,26 @@ export const REGIONS = [
   'us-east-1',
   'us-west-2',
   'eu-west-1',
+  'eu-west-2',
+  'eu-west-3',
+  'eu-east-2',
   'ap-south-1',
   'ap-southeast-2',
-  'eu-central-1'
-]
+  'eu-central-1',
+  'me-south-1',
+  'ap-east-1',
+  'ap-southeast-1',
+  'ap-northeast-1',
+  'ap-northeast-2',
+  'ap-northeast-3',
+  'ca-central-1',
+  'cn-north-1',
+  'cn-northwest-1',
+  'eu-north-1',
+  'sa-east-1',
+  ]
+
+
 
 const toArray = val => (val ? [].concat(val) : [])