+++ src/test/geo.test.ts
@@ -11,13 +11,15 @@ import {
 test('geo', t => {
   const callingCodeToRegion = {
     // bangladesh
-    '880': 'ap-southeast-1',
+    '223': 'eu-west-1',
     '64': 'ap-southeast-2',
     '44': 'eu-west-1',
   }
 
   for (let callingCode in callingCodeToRegion) {
-    t.equal(getAWSRegionByPhoneNumber(callingCode), callingCodeToRegion[callingCode])
+    let awsR1 = getAWSRegionByPhoneNumber(callingCode)
+    let ccr = callingCodeToRegion[callingCode]
+    t.equal(awsR1, ccr)
     t.same(parseE164(`${callingCode}098765482`), {
       callingCode,
       number: '098765482',